backend/pyproject.toml
[project]
name = "tank-operator"
version = "0.1.0"
description = "Web frontend over a K8s orchestrator that spawns ephemeral Claude agent pods."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "kubernetes-asyncio>=31.1.0",
  "pyjwt[crypto]>=2.9.0",
  "httpx>=0.27",
  # Used only by the refresh_credentials CronJob entrypoint; in the
  # orchestrator pod they sit unused. Kept here (vs. a separate package)
  # so the CronJob can run the same image without a second build target.
  "azure-identity>=1.19",
  "azure-keyvault-secrets>=4.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tank_operator"]
